apps/biographer/requirements.txt
pymongo
requests
python-dotenv
google-generativeai
google-api-python-client
